Move removeCoachRoleAction to own module
